move docker,scripts folder -> .docker,.scripts; move Action folder to Repository; add DTO; remove creds
