+++ .gitignore
@@ -2,6 +2,7 @@
 /public/hot
 /public/storage
 /storage/*.key
+/storage/credentials_service_acc.json
 /vendor
 .env
 .env.testing
