+++ app/Repository/AstrologerShow.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Action;
+namespace App\Repository;
 
 use App\Models\Astrologer;
